docs: swagger 폴더 분리
